refactor(middleware): use atomic files + docs
so middleware stays maintainable when adding more (like the upcoming file proxy middleware) and has a structure suitable to add tests/ to.
