+++ src/middleware/README.md
@@ -1,0 +1,3 @@
+# Middleware
+
+**[Astro supports middleware](https://docs.astro.build/en/guides/middleware/) that allows you to intercept requests and responses and inject behaviors dynamically every time a page or endpoint is about to be rendered.**
